Fixed responsebilities
I moved some responsebilities from the animal class to the cart class and vice versa so that it is now right.

I also made an Carnivore interface and an Herbivore interface so some sorting in core is easier if i wanted to add different animals who eat other animals.

Also made the Animal interface not dependent of a cart anymore by making it dependent on AnimalCollection interfaces